+++ contracts/test/TestExchange.sol
@@ -22,4 +22,10 @@ contract TestExchange is Exchange {
     function _blockTimestamp() internal view override returns (uint256) {
         return _testBlockTimestamp;
     }
+
+    // @dev max tick range = 887272 * 2
+    // @dev ref : https://github.com/Uniswap/v3-core/blob/main/contracts/libraries/TickMath.sol#L25
+    function _getMaxTickCrossedWithinBlockCap() internal pure override returns (uint24) {
+        return 1774544;
+    }
 }